+++ cmds_gen.py
@@ -25,10 +25,6 @@
         is_add_cmd
         run_prog
 
-        Disconnect (deprecated)
-        Is_Add_Cmd (deprecated)
-        Add_Cmd (deprecated)
-
 """
 
 ###############################################################################
@@ -219,90 +215,3 @@ def run_prog(cmd, **kwargs):
     else:
         P1 = subprocess.Popen(cmd)
         P1.wait()
-
-
-def Disconnect(*args):
-
-    """Function:  Disconnect (deprecated)
-
-    Description:  Disconnects a class database connection.  Will check to see
-        if an argument is an array; if so will loop on the array to disconnect
-        all connections.  Will require a disconnect method within the class.
-        The disconnect method will be particular to that class.
-
-    Arguments:
-        (input) *arg -> One or more connection instances.
-
-    """
-
-    for x in args:
-
-        if isinstance(x, list):
-            for y in x:
-                y.disconnect()
-
-        else:
-            x.disconnect()
-
-
-def Is_Add_Cmd(args_array, cmd, opt_arg_list, **kwargs):
-
-    """Function:  Is_Add_Cmd (deprecated)
-
-    Description:  Determine if any additional options need to be added to the
-        command line.
-
-    Arguments:
-        (input) args_array -> Array of command line options and values.
-        (input) cmd -> List array containing the program arguments.
-        (input) opt_arg_list -> Dictionary of additional options.
-        (input) **kwargs:
-            None
-        (output) cmd -> List array containing the program arguments.
-
-    """
-
-    for x in opt_arg_list:
-
-        # Is option in array and is set to True.
-        if x in args_array and args_array[x] and isinstance(args_array[x],
-                                                            bool):
-
-            if isinstance(opt_arg_list[x], list):
-
-                for y in opt_arg_list[x]:
-                    cmd = Add_Cmd(cmd, arg=y)
-
-            else:
-                cmd = Add_Cmd(cmd, arg=opt_arg_list[x])
-
-        elif x in args_array:
-            cmd = Add_Cmd(cmd, arg=opt_arg_list[x], val=args_array[x])
-
-    return cmd
-
-
-def Add_Cmd(cmd, **kwargs):
-
-    """Function:  Add_Cmd (deprecated)
-
-    Description:  Append name of argument and possibly value for the argument
-        to the command line list array.
-
-    Arguments:
-        (input) cmd -> List array containing the program setup.
-        (input) **kwargs:
-            arg -> Name of argument being added.
-            val -> Value for argument being added.
-        (output) cmd -> List array containing the program setup.
-
-    """
-
-    # Append before returning, appending on return does not make the change.
-    if "val" in kwargs:
-        cmd.append(kwargs["arg"] + kwargs["val"])
-
-    else:
-        cmd.append(kwargs["arg"])
-
-    return cmd